Neural Dynamics Model V2/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "swim-neural-model-v2"
version = "0.1.0"
description = "Pacinian-dominant computational neural dynamics model for ultrasonic haptics"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "scipy>=1.10",
  "h5py>=3.8",
  "pyyaml>=6.0",
  "matplotlib>=3.7",
  "numba>=0.58"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
